Creates Item namespace. Moves item-related enums into Item namespace. Renames Sandbox.ID to Sandbox.IDs. Moves all non-item sandbox enums into Sanbox.IDs namespace.
