+++ ParquetCSVImporter/Shims/CollectibleShim.cs
@@ -1,6 +1,6 @@
 using System;
 using ParquetClassLibrary;
-using ParquetClassLibrary.Sandbox.ID;
+using ParquetClassLibrary.Sandbox.IDs;
 using ParquetClassLibrary.Sandbox.Parquets;
 
 namespace ParquetCSVImporter.ClassMaps
