+++ ParquetClassLibrary/Crafting/CraftingRecipe.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Collections.Generic;
-using ParquetClassLibrary.Sandbox.ID;
 
 namespace ParquetClassLibrary.Crafting
 {
