+++ ParquetClassLibrary/Items/BiomeKey.cs
@@ -1,23 +1,35 @@
 ﻿using System;
 
-namespace ParquetClassLibrary.Sandbox
+namespace ParquetClassLibrary.Items
 {
     /// <summary>
-    /// Indicates a the level of a MapChunk or MapRegion.
+    /// Represents the subtype of <see cref="Items.ItemType.KeyItem"/> that grants safe access to 
+    /// the a <see cref="Biome"/>.  Indicates requirements a player character must meet to enter.
     /// </summary>
+    // TODO: This needs revision to represent all Key Item types.
     [Flags]
-    internal enum ElevationMask
+    public enum BiomeKey
     {
-        None,
-        BelowGround,
-        LevelGround,
-        AboveGround
+        None = 0,
+        Knapsack,
+        Sunscreen,
+        Boots,
+        Canteen,
+        BugSpray,
+        WarmGear,
+        Lantern,
+        Rope,
+        ClimbingKit,
+        GasMask,
+        Charm,
+        PearlyFeather,
+        CharredMask,
     }
 
     /// <summary>
-    /// Convenience extension methods for concise coding when working with <see cref="ElevationMask"/> instances.
+    /// Convenience extension methods for concise coding when working with <see cref="BiomeKey"/> instances.
     /// </summary>
-    internal static class ElevationMaskSelectionExtensions
+    internal static class BiomeKeyExtensions
     {
         /// <summary>
         /// Sets the given flag in the specified variable.
@@ -25,7 +37,7 @@ namespace ParquetClassLibrary.Sandbox
         /// <param name="in_enumVariable">The enum variable under consideration.</param>
         /// <param name="in_flagToTest">The flag to set.</param>
         /// <returns>The variable with the flag set.</returns>
-        public static ElevationMask Set(this ref ElevationMask in_enumVariable, ElevationMask in_flagToTest)
+        public static BiomeKey Set(this ref BiomeKey in_enumVariable, BiomeKey in_flagToTest)
         {
             return in_enumVariable = in_enumVariable | in_flagToTest;
         }
@@ -36,7 +48,7 @@ namespace ParquetClassLibrary.Sandbox
         /// <param name="in_enumVariable">The enum variable under consideration.</param>
         /// <param name="in_flagToTest">The flag to clear.</param>
         /// <returns>The variable with the flag cleared.</returns>
-        public static ElevationMask Clear(this ref ElevationMask in_enumVariable, ElevationMask in_flagToTest)
+        public static BiomeKey Clear(this ref BiomeKey in_enumVariable, BiomeKey in_flagToTest)
         {
             return in_enumVariable = in_enumVariable & ~in_flagToTest;
         }
@@ -47,7 +59,7 @@ namespace ParquetClassLibrary.Sandbox
         /// <param name="in_enumVariable">The enum variable under consideration.</param>
         /// <param name="in_flagToTest">The flag to test.</param>
         /// <returns><c>true</c>, if at least this flag is set, <c>false</c> otherwise.</returns>
-        public static bool IsSet(this ElevationMask in_enumVariable, ElevationMask in_flagToTest)
+        public static bool IsSet(this BiomeKey in_enumVariable, BiomeKey in_flagToTest)
         {
             return (in_enumVariable & in_flagToTest) == in_flagToTest;
         }
@@ -58,9 +70,9 @@ namespace ParquetClassLibrary.Sandbox
         /// <param name="in_enumVariable">The enum variable under consideration.</param>
         /// <param name="in_flagToTest">The flag to test.</param>
         /// <returns><c>true</c>, if at least this flag is unset, <c>false</c> otherwise.</returns>
-        public static bool IsNotSet(this ElevationMask in_enumVariable, ElevationMask in_flagToTest)
+        public static bool IsNotSet(this BiomeKey in_enumVariable, BiomeKey in_flagToTest)
         {
-            return (in_enumVariable & ~in_flagToTest) == ElevationMask.None;
+            return (in_enumVariable & ~in_flagToTest) == BiomeKey.None;
         }
 
         /// <summary>
@@ -70,7 +82,7 @@ namespace ParquetClassLibrary.Sandbox
         /// <param name="in_flagToTest">The flag to set or clear.</param>
         /// <param name="in_state">If <c>true</c>, the flag will be set; otherwise it will be cleared.</param>
         /// <returns>The variable with the flag modified.</returns>
-        public static ElevationMask SetTo(this ref ElevationMask in_enumVariable, ElevationMask in_flagToTest, bool in_state)
+        public static BiomeKey SetTo(this ref BiomeKey in_enumVariable, BiomeKey in_flagToTest, bool in_state)
         {
             return in_enumVariable = in_state
                 ? in_enumVariable.Set(in_flagToTest)
