+++ ParquetClassLibrary/Items/GatheringTools.cs
@@ -1,4 +1,4 @@
-﻿namespace ParquetClassLibrary.Sandbox.ID
+﻿namespace ParquetClassLibrary.Items
 {
     /// <summary>
     /// IDs for tools that Characters can use to gather Blocks.
