+++ ParquetClassLibrary/Sandbox/IDs/CollectionEffect.cs
@@ -1,4 +1,4 @@
-﻿namespace ParquetClassLibrary.Sandbox.ID
+﻿namespace ParquetClassLibrary.Sandbox.IDs
 {
     /// <summary>
     /// IDs for effects that can happen when a character encounters a Collectible.
