+++ ParquetClassLibrary/Sandbox/IDs/GatheringEffect.cs
@@ -1,4 +1,4 @@
-namespace ParquetClassLibrary.Sandbox.ID
+namespace ParquetClassLibrary.Sandbox.IDs
 {
     /// <summary>
     /// IDs for effects that can happen when a character gathers a Block.
