+++ ParquetClassLibrary/Sandbox/Parquets/Collectible.cs
@@ -1,6 +1,6 @@
 using System;
 using Newtonsoft.Json;
-using ParquetClassLibrary.Sandbox.ID;
+using ParquetClassLibrary.Sandbox.IDs;
 using ParquetClassLibrary.Utilities;
 
 namespace ParquetClassLibrary.Sandbox.Parquets
