+++ ParquetUnitTests/Sandbox/ChunkTypeUnitTest.cs
@@ -1,5 +1,5 @@
 ﻿using System;
-using ParquetClassLibrary.Sandbox;
+using ParquetClassLibrary.Sandbox.IDs;
 using Xunit;
 
 namespace ParquetUnitTests.Sandbox
