+++ ParquetUnitTests/Sandbox/MapChunkGridUnitTest.cs
@@ -1,5 +1,6 @@
 using System;
 using ParquetClassLibrary.Sandbox;
+using ParquetClassLibrary.Sandbox.IDs;
 using ParquetClassLibrary.Stubs;
 using Xunit;
