+++ ParquetUnitTests/Sandbox/MapRegionUnitTest.cs
@@ -1,6 +1,7 @@
 using System;
 using ParquetClassLibrary;
 using ParquetClassLibrary.Sandbox;
+using ParquetClassLibrary.Sandbox.IDs;
 using ParquetClassLibrary.Sandbox.SpecialPoints;
 using ParquetClassLibrary.Stubs;
 using ParquetUnitTests.Sandbox.Parquets;
